Make it possible to choose between slime and sly
Make it possible for the app to decide to load the slynk or swank
backend.

This commit also adapts the precompile section in install.lisp to
preload both slynk and swank, so it starts up faster later. The rest
of the precompile stuff will be moved to application build time to
make it more versatile.

It introduces two new variables that the application can use at run
time: DEV_BACKEND chooses between sly and slime backends; If it is
unset, neither will be loaded. DEV_BACKEND_PORT can be used to choose
another slynk/swank port if the application requires it. It will
only make a difference if a DEV_BACKEND is used. The default port is
4005.