Install up-to-date sbcl-1.4.12 from custom RPM
The sbcl-1.4.0 contained in the epel is too old, some stuff in
quicklisp or other distributions do not compile with the ASDF/UIOP it
contains. Let's fetch an up-to-date sbcl rpm from an external source.
This commit also splits up the build process so that future sbcl
updates can be performed faster.