When using slynk, do not use a dedicated output stream
The default is t, causing the connection to hang when using OpenShift
port-forwarding. To use OpenShift port-forwarding, we need to do the
same as when tunneling via SSH, because there is only one port to use,
and therefore set slynk's dedicated output stream to nil. For more
information about this, see the section in the fine sly manual
http://joaotavora.github.io/sly/#Setting-up-the-Lisp-image