Lower sb-ext:*exit-timeout* from 60 to 15 seconds
This is the timeout SBCL waits for threads to terminate. For
OpenShift, we need to lower this from the default 60 seconds so that
containers get shutdown in time for deployment of new containers. The
application should ensure a proper shutdown routine is called by
adding a function to sb-ext:*exit-hooks* that gets called on sbcl
exit. If the new value (15 seconds) is too low for some reason, the
application can override this in its run script.