Update SBCL to 1.4.13 and use upstream source
For now and also for future releases, download the precompiled
sbcl-1.4.13 from sbcl.org (actually from a SourceForge mirror) instead
of using an ALTLinux package and install it in the default location
in /usr/local.

Also split up the installation of and rearrange the order of library
packages and lisp packages to make better use of the docker build
caching mechanism. This also improves the clarity of the Dockerfile a
bit.